+++ packages/dataviews/CHANGELOG.md
@@ -18,6 +18,10 @@
 - Add support for the `Edit` control on the date field type ([#70836](https://github.com/WordPress/gutenberg/pull/70836)).
 - Add support for responsive image attributes in media fields by adding a `config` prop to `mediaField.render`, containing a `sizes` string. Also simplifies grid logic and fixes imprecisions in grid resizing by changing the observed container. ([#70493](https://github.com/WordPress/gutenberg/pull/70493)).
 
+### Enhancements
+
+- Make the media item clickable along the title ([#70985](https://github.com/WordPress/gutenberg/pull/70985)).
+
 ## 5.0.0 (2025-07-23)
 
 ### Bug Fixes